@@ -12,6 +12,7 @@
     "firebaseui": "^3.4.1",
     "ganache-cli": "^6.1.8",
     "newsapi": "^2.4.0",
+    "prop-types": "^15.6.2",
     "react": "^16.5.2",
     "react-chartjs-2": "^2.7.4",
     "react-dom": "^16.5.2",
@@ -35,7 +36,7 @@
   },
   "lint": "eslint src/*",
   "devDependencies": {
-    "eslint": "^5.3.0",
+    "eslint": "^5.6.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
