@@ -23,7 +23,8 @@ class SingleDiscourse extends Component {
       return (
         <div>
           <div className='discourse-container'>
-            <h3>publicDiscourse: {title}</h3>
+          {/* This Thing renders terribly we need to delete it. */}
+            {/* <h3>publicDiscourse: {title}</h3> */}
             <div className='single-room'>
               <div className='iframe-container'>
                 <iframe src={url} />
